requirements.txt
customtkinter
openai-whisper
moviepy==1.0.3
torch
numpy
Pillow
